Add .env.example and update application.yml for environment variable configuration

+++ spring-expert/libraryapi/src/main/java/com/study/libraryapi/model/Livro.java
@@ -1,5 +1,6 @@
 package com.study.libraryapi.model;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 import java.util.UUID;
 
@@ -39,8 +40,8 @@ public class Livro {
     private GeneroLivro genero;
 
     @Column(name = "preco", precision = 18, scale = 2)
-    private Double preco;
-    // private BigDecimal preco;
+    private BigDecimal preco;
+    // private Double preco;
 
     @ManyToOne
     @JoinColumn(name = "id_autor")
